refactor: auth 코드 이동
